Summary notebook  and ODE_solver  update
I have wrote a summary notebook of what I learned. I used the ODE_solver as a demonstration.

The ODE_Solver is able to 'accurately' solve the first 3 time steps (as accurate as we can get with the approximations we imposed).  The 4th time slice and up requires more work. (inaccurate solutions).